Refactor PySide6 main window into modular packages
